+++ eiam-20211201/alibabacloud_eiam20211201/models/_obtain_credential_response_body.py
@@ -48,6 +48,7 @@ class ObtainCredentialResponseBodyCredential(DaraModel):
         create_time: int = None,
         credential_content: main_models.ObtainCredentialResponseBodyCredentialCredentialContent = None,
         credential_creation_type: str = None,
+        credential_external_id: str = None,
         credential_id: str = None,
         credential_identifier: str = None,
         credential_name: str = None,
@@ -68,6 +69,7 @@ class ObtainCredentialResponseBodyCredential(DaraModel):
         self.credential_content = credential_content
         # 凭据的创建类型。
         self.credential_creation_type = credential_creation_type
+        self.credential_external_id = credential_external_id
         # 凭据ID。
         self.credential_id = credential_id
         # 凭据标识
@@ -110,6 +112,9 @@ class ObtainCredentialResponseBodyCredential(DaraModel):
 
         if self.credential_creation_type is not None:
             result['CredentialCreationType'] = self.credential_creation_type
+
+        if self.credential_external_id is not None:
+            result['CredentialExternalId'] = self.credential_external_id
 
         if self.credential_id is not None:
             result['CredentialId'] = self.credential_id
@@ -163,6 +168,9 @@ class ObtainCredentialResponseBodyCredential(DaraModel):
 
         if m.get('CredentialCreationType') is not None:
             self.credential_creation_type = m.get('CredentialCreationType')
+
+        if m.get('CredentialExternalId') is not None:
+            self.credential_external_id = m.get('CredentialExternalId')
 
         if m.get('CredentialId') is not None:
             self.credential_id = m.get('CredentialId')