+++ eiam-20211201/ChangeLog.md
@@ -1,3 +1,11 @@
+2026-05-13 Version: 2.15.9
+- Update API CreateCredential: add request parameters CredentialExternalId.
+- Update API GetCredential: add response parameters Body.Credential.CredentialExternalId.
+- Update API ListCredentials: add request parameters CredentialExternalIds.
+- Update API ListCredentials: add response parameters Body.Credentials.$.CredentialExternalId.
+- Update API ObtainCredential: add response parameters Body.Credential.CredentialExternalId.
+
+
 2026-04-28 Version: 2.15.8
 - Update API GetInstanceControlConfiguration: add request parameters ElementName.